extract all StrEnums to new matbench_discovery/enums.py
